Move user endpoints into routes & controllers
