+++ src/server/models/user.js
@@ -6,7 +6,7 @@ var UserSchema = new mongoose.Schema(
         name: {type: String, required: true},
         hashedPassword: {type: String, required: true},
         salt: {type: String, required: true},
-        userType: {type: String, enum: ['customer', 'seller', 'admin'], required: true},
+        userType: {type: String, enum: ['Customer', 'Seller', 'Admin'], required: true},
         address: String,
         creditCard: String
     }
